removing response_handler from Alter query
Added a lot of complexity, and potential performance downsides, to migrations which add columns. The original intention was if a default is a callable, the callable would be executed to create a unique value for each existing row, after the new column was added. This is useful when say the new value was a random number, uuid, or such. The UUID use case is now covered using the `columns.defaults.uuid.UUID4` class. In situations where the callable always generates the same value, the existing approach was very wasteful. This might be revisited in the future, if there are edge cases which require this behaviour - and if so, syntax will be added to support 'unique callables' vs 'static callables'.