Add unique global names to serialisation (#306)
Adds a UniqueGlobalNames class to serialisation that stores
global names as class attributes. These attributes are meant
to be used during serialisation to prevent conflicting global
names in the generated code.

As a result of runtime conflict prevention the following errors
are raises while running the test suite:
- `piccolo.apps.migrations.auto.serialisation.UniqueGlobalNameConflictError: Import 'Decimal' could conflict with global name`
- `piccolo.apps.migrations.auto.serialisation.UniqueGlobalNameConflictError: Import 'UUID' could conflict with global name`

These error match the errors found in branch
/0scarB/piccolo/tree/fix-auto-migration-import-conflict