change `run_sync` (#539)
* change `run_sync`

* simplify `run_sync` - call `asyncio.run` first

Trying `asyncio.run` first, as in 99% of cases this will work.

Also, using asyncio.run in a thread when running in an event loop.

* Create test-strict.sh

This runs the tests in Python's development mode, which gives lots of useful information - for example, about deprecated features, and asyncio warnings.

* close connection if exception is raised by asyncpg

When running the tests in development mode, I realised that there were situations where the connection wasn't being closed (i.e. if an exception was raised by asyncpg due to a Postgres error).

* added `create_db_tables` and `drop_db_tables`

These are async first. They replace `create_tables` and `drop_tables` which were sync first, and had no async equivalent.