Altering `Ondelete` and `OnUpdate` with `MigrationManager` (#1183)
* fix OnDelete and OnUpdate in migrations

* fix test docstring

* remove for loop

* refactor `add_foreign_key_constraint`

If a `ForeignKey` is passed in, we can get most of the info we need from it, rather than requiring additional params to be passed in.

* add `get_fk_constraint_name`

* rename pk constraint on table renaming in migrations

* `RenamePkConstraint` -> `RenameConstraint`

* skip renaming pk constraint in SQLite

* revert changes

* update param names in `rename_constraint` method

* add docstring

* add `get_fk_constraint_rules` function

So the logic for getting constraints is centralised.

* add TODO for supporting SQLite

* add TODO for moving logic from `generate.py` to `constraints.py`

* remove redundant lookup in `generate.py`

We should refactor a bunch of this - I came across this dict lookup, which isn't required because we can just call the enum directly.

---------

Co-authored-by: Daniel Townsend <dan@dantownsend.co.uk>